Reorganizing file locations for clarity
