+++ server/routes/contact.js
@@ -1,17 +1,15 @@
-import express from 'express'
+import { Router } from 'express'
 import consola from 'consola'
 
 // Contact Model
 import Contact from '../models/contact'
 
-const app = express()
+const router = Router()
 
 // @route   GET api/contact/contacts
 // @desc    Gets all contacts
 // @access  Public
-
-// eslint-disable-next-line space-before-function-paren
-app.get('/contacts', async (req, res) => {
+router.get('/contacts', async(req, res) => {
   try {
     const contacts = await Contact.find()
 
@@ -32,7 +30,7 @@ app.get('/contacts', async (req, res) => {
 // @access  Public
 
 // eslint-disable-next-line space-before-function-paren
-app.get('/contact', async (req, res) => {
+router.get('/contact', async (req, res) => {
   try {
     const contact = await Contact.findById(req.query.id)
 
@@ -51,9 +49,7 @@ app.get('/contact', async (req, res) => {
 // @route   POST api/contact/contact/addContact
 // @desc    Post creates a new contact
 // @access  Public
-
-// eslint-disable-next-line space-before-function-paren
-app.post('/addContact', async (req, res) => {
+router.post('/addContact', async(req, res) => {
   let errMessage = ''
 
   try {
@@ -93,4 +89,4 @@ app.post('/addContact', async (req, res) => {
   }
 })
 
-export default app
+export default router
